requirements.txt
python-dotenv==0.17.1
python-telegram-bot==20.3
PyPDF2==3.0.1
reportlab==4.0.4
stripe==8.7.0
